requirements.txt
flask
psycopg2-binary
Flask-SQLAlchemy
Flask-CORS
flask-marshmallow
marshmallow-SQLAlchemy